Desenvolvimento da função de ocupação e desocupação do quarto

+++ backend/src/controller/Room.ts
@@ -1,4 +1,5 @@
 import { Request, Response } from "express";
+import { createHistory, freeBedroom, isBedroomInUse } from "../services/BedroomHistoryService";
 import { hasBedroomNumber, registerBedroom, getAll, deleteRoom, getRoom } from "../services/BedroomService";
 
 export async function index(req: Request, res: Response) {
@@ -12,11 +13,36 @@ export async function get(req: Request, res: Response) {
 export async function deleteRequest(req: Request, res: Response) {
     const id = getRoomId(req);
     if (id < 1)
-        return res.status(404).json({ message: 'O quarto é inválido.' });
+        return res.status(400).json({ message: 'O quarto é inválido.' });
+
+    if (await isBedroomInUse(id))
+        return res.status(400).json({ message: 'Não é possível apagar um quarto enquanto ele está em uso.' });
 
     await deleteRoom(id);
     res.json({ message: 'Ok' });
 };
+
+export async function freeRoom(req: Request, res: Response) {
+    const id = Number(req.params.id);
+    if (id < 1)
+        return res.status(400).json({ message: 'O quarto é inválido.' });
+
+    await freeBedroom(id);
+    res.json({ message: 'Ok.' });
+}
+
+export async function setCustommer(req: Request, res: Response) {
+    const id = Number(req.params.id);
+    if (id < 1)
+        return res.status(400).json({ message: 'O quarto é inválido.' });
+
+    const idCustomer = Number(req.body.idCustomer);
+    if (id < 1)
+        return res.status(400).json({ message: 'O cliente é inválido.' });
+
+    await createHistory(req.user?.id as any, idCustomer, id);
+    res.json({ message: 'Ok.' });
+}
 
 export async function register(req: Request, res: Response) {
     const roomNumber = getRoomNumber(req);
